requirements.txt
-i https://pypi.org/simple
altair==5.2.0; python_version >= '3.8'
attrs==23.2.0; python_version >= '3.7'
blinker==1.7.0; python_version >= '3.8'
cachetools==5.3.2; python_version >= '3.7'
certifi==2024.2.2; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
gitdb==4.0.11; python_version >= '3.7'
gitpython==3.1.41; python_version >= '3.7'
idna==3.6; python_version >= '3.5'
importlib-metadata==7.0.1; python_version >= '3.8'
jinja2==3.1.3; python_version >= '3.7'
jsonschema==4.21.1; python_version >= '3.8'
jsonschema-specifications==2023.12.1; python_version >= '3.8'
markdown-it-py==3.0.0; python_version >= '3.8'
markupsafe==2.1.5; python_version >= '3.7'
mdurl==0.1.2; python_version >= '3.7'
numpy==1.26.4; python_version == '3.11'
packaging==23.2; python_version >= '3.7'
pandas==2.2.0; python_version >= '3.9'
pillow==10.2.0; python_version >= '3.8'
protobuf==4.25.2; python_version >= '3.8'
pyarrow==15.0.0; python_version >= '3.8'
pydeck==0.8.1b0; python_version >= '3.7'
pygments==2.17.2; python_version >= '3.7'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.1; python_version >= '3.8'
pytz==2024.1
referencing==0.33.0; python_version >= '3.8'
requests==2.31.0; python_version >= '3.7'
rich==13.7.0; python_full_version >= '3.7.0'
rpds-py==0.17.1; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
smmap==5.0.1; python_version >= '3.7'
streamlit==1.31.0; python_version >= '3.8' and python_full_version != '3.9.7'
tenacity==8.2.3; python_version >= '3.7'
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
toolz==0.12.1; python_version >= '3.7'
tornado==6.4; python_version >= '3.8'
typing-extensions==4.9.0; python_version >= '3.8'
tzdata==2023.4; python_version >= '2'
tzlocal==5.2; python_version >= '3.8'
urllib3==2.2.0; python_version >= '3.8'
validators==0.22.0; python_version >= '3.8'
zipp==3.17.0; python_version >= '3.8'
